refactor(coordinator):  address some minor rounding issues

+++ coordinator/ethereum/gas-pricing/dynamic-cap/src/main/kotlin/net/consensys/linea/ethereum/gaspricing/dynamiccap/GasPriceCapProviderImplV2.kt
@@ -55,7 +55,7 @@ class GasPriceCapProviderImplV2(
     val minNumOfFeeHistoriesNeeded =
       config.gasFeePercentileWindowInBlocks
         .minusCoercingUnderflow(config.gasFeePercentileWindowLeewayInBlocks)
-        .toInt()
+        .toLong()
 
     val numOfValidFeeHistories = feeHistoriesRepository.getCachedNumOfFeeHistoriesFromBlockNumber()
     val hasEnoughData = numOfValidFeeHistories >= minNumOfFeeHistoriesNeeded
